+++ HTML/html_document.cpp
@@ -7,9 +7,9 @@
 #include "../PThread/pthread_internal.hpp"
 #include <new>
 
-thread_local uint32_t html_document::_last_error = static_cast<uint32_t>(FT_ERR_SUCCESS);
+thread_local int32_t html_document::_last_error = FT_ERR_SUCCESS;
 
-uint32_t html_document::set_error(uint32_t error_code) noexcept
+int32_t html_document::set_error(int32_t error_code) noexcept
 {
     html_document::_last_error = error_code;
     return (error_code);
