+++ Parser/parser_dom_node.cpp
@@ -364,7 +364,7 @@ int32_t ft_dom_node::add_child(ft_dom_node *child) noexcept
         return (lock_error);
     this->_children.push_back(child);
     (void)this->unlock_internal(lock_acquired);
-    return (static_cast<int32_t>(this->_children.get_error()));
+    return (this->_children.get_error());
 }
 
 const ft_vector<ft_dom_node*> &ft_dom_node::get_children() const noexcept
@@ -401,11 +401,11 @@ int32_t ft_dom_node::add_attribute(const ft_string &key, const ft_string &value)
     if (this->_attribute_keys.get_error() != FT_ERR_SUCCESS)
     {
         (void)this->unlock_internal(lock_acquired);
-        return (static_cast<int32_t>(this->_attribute_keys.get_error()));
+        return (this->_attribute_keys.get_error());
     }
     this->_attribute_values.push_back(value);
     (void)this->unlock_internal(lock_acquired);
-    return (static_cast<int32_t>(this->_attribute_values.get_error()));
+    return (this->_attribute_values.get_error());
 }
 
 int32_t ft_dom_node::add_attribute(const char *key, const char *value) noexcept
