+++ Test/Test/test_template_circular_buffer_move.cpp
@@ -36,12 +36,12 @@ FT_TEST(test_ft_circular_buffer_dynamic_reset_after_errors)
     buffer_instance.push(1);
     buffer_instance.push(2);
     buffer_instance.push(3); // triggers full error
-    FT_ASSERT_EQ(FT_ERR_FULL, static_cast<int32_t>(buffer_instance.get_error()));
+    FT_ASSERT_EQ(FT_ERR_FULL, buffer_instance.get_error());
     buffer_instance.clear();
     FT_ASSERT_EQ(true, buffer_instance.is_empty());
     FT_ASSERT_EQ(FT_ERR_SUCCESS, buffer_instance.get_error());
     buffer_instance.pop(); // empty underflow
-    FT_ASSERT_EQ(FT_ERR_EMPTY, static_cast<int32_t>(buffer_instance.get_error()));
+    FT_ASSERT_EQ(FT_ERR_EMPTY, buffer_instance.get_error());
     buffer_instance.push(4);
     FT_ASSERT_EQ(FT_ERR_SUCCESS, buffer_instance.get_error());
     FT_ASSERT_EQ(FT_ERR_SUCCESS, buffer_instance.destroy());
