+++ Test/Test/test_thread_pool.cpp
@@ -22,7 +22,7 @@ FT_TEST(test_thread_pool_resets_error_status)
         execution_count.store(-1);
         return ;
     });
-    FT_ASSERT_EQ(FT_ERR_NO_MEMORY, static_cast<int32_t>(pool_instance.get_error()));
+    FT_ASSERT_EQ(FT_ERR_NO_MEMORY, pool_instance.get_error());
     cma_set_alloc_limit(0);
     pool_instance.submit([&execution_count]()
     {
